+++ src/main/java/com/banking/springbootbanking/service/TransactionService.java
@@ -11,17 +11,17 @@ public interface TransactionService {
 
     TransactionDTO getTransactionById(Long id);
 
-    TransactionDTO accountWithdraw(String number, Long amount, CurrencyType currency);
+    TransactionDTO accountWithdraw(String number, Float amount, CurrencyType currency);
 
-    TransactionDTO cardWithdraw(String number, Long amount, CurrencyType currency);
+    TransactionDTO cardWithdraw(String number, Float amount, CurrencyType currency);
 
-    TransactionDTO accountDeposit(String number, Long amount, CurrencyType currency);
+    TransactionDTO accountDeposit(String number, Float amount, CurrencyType currency);
 
-    TransactionDTO cardDeposit(String number, Long amount, CurrencyType currency);
+    TransactionDTO cardDeposit(String number, Float amount, CurrencyType currency);
 
-    TransactionDTO accountTransfer(String senderNumber, String recipientNumber, Long amount, String description, CurrencyType currency);
+    TransactionDTO accountTransfer(String senderNumber, String recipientNumber, Float amount, String description, CurrencyType currency);
 
-    TransactionDTO cardTransfer(String senderNumber, String recipientNumber, Long amount, String description, CurrencyType currency);
+    TransactionDTO cardTransfer(String senderNumber, String recipientNumber, Float amount, String description, CurrencyType currency);
 
     List<TransactionDTO> getAllTransactions();
 }
